+++ src/App.jsx
@@ -1,0 +1,8 @@
+import Footer from "./components/footer/Footer.jsx";
+import "./App.css";
+
+function App() {
+  return <Footer />;
+}
+
+export default App;
