+++ src/components/header/NavBar.jsx
@@ -1,0 +1,5 @@
+function NavBar(){
+  
+}
+
+export default NavBar
